yacc: Update enumerations definition
The current model for enumerated types only accepts construction via a
NamedNumberList. This doesn't allow for much of the expressiveness encoded
in X.680 C20.XX on the enumerated types.

This patch brings about a new yacc type 'Enumerations', along with it's
construction type 'Enumeration' which can be either a NamedNumber or a simple
identifier.

Signed-off-by: Aaron Conole <aconole@bytheb.org>
fixes: #7