Makefile.am: Remove nonstandard functions
The .asn1.c and .asn1.cpp targets use GNU Make specific functions. This is
non-portable, and throws warnings as such when running the included boot.sh
to generate the configure scripts and Makefiles.

This change uses the newly introduced '-mo' facility to shorten the rules
and remove the non-standard functions.

Signed-off-by: Aaron Conole <aconole@bytheb.org>