gen-buf: Correct the genbuf macros and functions
First, the BufPeekSegFcn prototype incorrectly reported a pointer for
the lenPtr argument. This commit corrects that by setting the value to a
pointer type. It just so happens that the width of unsigned long is
enough to hold a pointer on most systems.

Second, change the macros to be proper static inlines, meaning the
warnings systems will properly catch calling issues.

Signed-off-by: Aaron Conole <aconole@bytheb.org>