min-buf: Add a min-buf facility
The previous incarnation of minbuf used the old-style macro
settings. This became invalid quite some time ago, but the minbuf was
never updated to use the indirection layer minbuf.

This commit introduces a proper minbuf facility. A future commit will
enable the examples so that a proper first-cut will be available.

Signed-off-by: Aaron Conole <aconole@bytheb.org>