c-lib/inc/asn-len.h: Change indefinite len to ul
Indefinite length is currently spelled out as a signed value, but the
typical use case (using it with an AsnLen type) is an unsigned type.

It doesn't make sense for length to be negative in either ASN.1 or in
the eSNACC suite, so just make it unsigned and deal with it.

Signed-off-by: Aaron Conole <aconole@redhat.com>