snacc.h: Add the ability to indicate unused vars
Some variables are part of an abstract interface but will not be used
by the specific implementation. This may due to API/ABI compliance, or
because more efficient sub-implementations can be done.

This requires adding a per-compiler 'unused' attribute
specifier. Additionally, this commit adds an example in the min-buf case.

Signed-off-by: Aaron Conole <aconole@redhat.com>