c-lib: Cleanup unused warnings
During compilation, a number of warnings are triggered by some compilers that
variables are being self-referenced.  Removing this reference would cause
a different warning - an 'unused variable' warning.  With this change, we
switch to indicating that the variable is not used.

At some point in the future, we should break API and ABI compat and remove
these.

Suggested-by: Arpad Tigyi <tigyi.arpad@gmail.com>
Signed-off-by: Aaron Conole <aconole@bytheb.org>