c-lib/asn-int: Shift with defined behavior
Left-shifting with a negative value is not well-defined in C.  This means
that any compiler is free to implement the consequences of that shift in
whatever manner it sees fit, including doing something wrong.

Since we really just want to set the upper bits to indicate the value is
negative, we instead start with a bitflipped 0.

Suggested-by: Arpad Tigyi <tigyi.arpad@gmail.com>
Signed-off-by: Aaron Conole <aconole@bytheb.org>