cxx-lib: remove vdathread support
VDAThreads was an unused and psuedo-proprietary method for coordinating
multiple threads to share access to memory and buffers.  After
discussion with Pierce Leonberger of BAE Systems, I don't see a need to
continue supporting it.  Modern environments offer thread support, and such
support is baked into the languages (c++11, c11, etc).  Even if it weren't
there is no reason for eSNACC to provide it.

Signed-off-by: Aaron Conole <aconole@bytheb.org>
Reviewed-by: Sagara Wickramasekara <sagaraw@gmail.com>