cxx-lib: do not use exceptions in performance critical operators
In asn-buf.cpp the operator<() and the operator==() methods
suffer from a performance issue, they are running in a loop until
an exception is not thrown. This is a very bad idea, and very
inefficient way of error handling. I have done some performance
tests with gproof. The performance has improved about 4 times better,
the code runs faster after my fix. I removed the exception handling from
operator<() and the operator==() and replaced them to a
boolean function, that can signal if an exception would be thrown.

Signed-off-by: László Kovács <sabbath24@gmail.com>
Acked-by: Aaron Conole <aconole@bytheb.org>