cross-build: disable asn rose
When cross-compiling the system, it is not possible for the esnacc binary to
correctly execute on the build system (well, not always).  There is a
method to combat this, which is the approach gcc takes, to generate a stage
one version of the compiler and then execute that to generate the final
tools (stage 2).  That represents quite a bit of work.

On the other hand, people are using esnacc *now* and as Marty Galligan
reports at http://mail.esnacc.org/pipermail/dev/2017-July/000414.html there
can be problems cross building, especially related to this.

This commit detects that a cross-compilation is occuring and disables the
generation of the asn.1 ROSE support.  A future commit can enable this
when it becomes necessary.

Acked-by: Marty Galligan <m_galligan@comcast.net>
Reported-by: Marty Galligan <m_galligan@comcast.net>
Signed-off-by: Aaron Conole <aconole@bytheb.org>